requirements.txt
fastapi
uvicorn
langchain-openai
langchain-community
faiss-cpu
sentence-transformers
datasets
python-dotenv
numpy